The Nquuens Algorithm with Explaination and Lesser Complexity
