+++ backtracking/n-queens/Python/Sairam-K26.txt
@@ -1,0 +1,21 @@
+The N-Queens problem is a classic puzzle where the goal is to place N chess queens on an NxN chessboard so that no two queens attack each other. In other words, no two queens can share the same row, column, or diagonal.
+
+Here's an example of how we can visualize the N-Queens problem in a game format with a 4x4 board and 4 queens:
+  a b c d
+1 Q . . .
+2 . . Q .
+3 . Q . .
+4 . . . Q
+
+Each square on the board represents a position where a queen can be placed. In this example, we have placed four queens on the board so that no two queens share the same row, column, or diagonal.
+
+Now, let's say you have a specific input, such as a 5x5 board and 5 queens. Here's how we can visualize the game with this input:
+  a b c d e
+1 . . Q . .
+2 Q . . . .
+3 . . . . Q
+4 . Q . . .
+5 . . . Q .
+
+
+
